Fixed a bug. Bug details: I read from InputStream so I try to fill array of `n` bytes, then the read operation is blocked until it fills at list one byte but there is 1=<=x<=n bytes to read, then it will read only `x` bytes. This is a bug because I expect to read `n` bytes or to get exception because the peer closed the connection.
Why is this a bug: Because in the next read operation, I will start reading the end of the message before, instead of reading the new message.

Fix: I'm usning DataInputStream instead of InputStream because it have a method: readFully which is blocking operation until it read exactly `n` bytes or the peer closed the connection so we get an exception.